+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for Python 3.9.
 - Short `-n` option for `--name` option in `pybricksdev run`.
+- Option to set hub name when flashing firmware.
 
 ### Changed
 - Update to Bleak v0.12.0.